requirements.txt
groq
python-dotenv
speechrecognition
pyttsx3
selenium
googlesearch-python
requests
numpy
pydub
sounddevice
nltk
yfinance
newsapi-python
youtube-search-python
google-cloud-texttospeech
PyQt6>=6.6